Fix type stability

+++ src/AIAPC2020/policy.jl
@@ -24,8 +24,8 @@ AIAPCPolicy(env::AIAPCEnv) = MultiAgentPolicy(
             ),
         Trajectory(CircularArraySARTTraces(;
         capacity=3,
-        state=Int16 => (225,),
-        action=Int8 => (15,),
+        state=Int64 => (225,),
+        action=Int64 => (15,),
         reward=Float32 => (),
         terminal=Bool => ()
         ), DummySampler()),
